requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
json5==0.9.25
psycopg2==2.9.9
python-dotenv
cryptocode